agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hexascan-agent"
version = "1.0.0"
description = "HexaScan monitoring agent for client servers"
authors = [{name = "HexaScan"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "cryptography>=41.0.0",
]

[project.scripts]
hexascan-agent = "hexascan_agent.agent:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hexascan_agent*"]
